Move model directory
