+++ src/backend/tests/expunger/test_expunger.py
@@ -3,7 +3,7 @@ import unittest
 from datetime import date
 from dateutil.relativedelta import relativedelta
 from expungeservice.expunger.expunger import Expunger
-from expungeservice.crawler.models.disposition import Disposition
+from expungeservice.models.disposition import Disposition
 from tests.factories.case_factory import CaseFactory
 from tests.factories.charge_factory import ChargeFactory
