+++ src/backend/tests/expunger/test_type_analyzer_convicted_charges.py
@@ -2,9 +2,8 @@ import unittest
 
 from datetime import datetime, timedelta
 from expungeservice.expunger.analyzers.type_analyzer import TypeAnalyzer
-from tests.factories.case_factory import CaseFactory
 from tests.factories.charge_factory import ChargeFactory
-from expungeservice.crawler.models.charge import Disposition
+from expungeservice.models.charge import Disposition
 
 
 class TestSingleChargeConvictions(unittest.TestCase):
