+++ src/backend/tests/factories/case_factory.py
@@ -1,4 +1,4 @@
-from expungeservice.crawler.models.case import Case
+from expungeservice.models.case import Case
 
 
 class CaseFactory:
